Update patch set 3
Patch Set 3:

yeah, bird and I were talking about that yesterday.  I was able to salvage a possible script we could use that I had already written when trying to do that with GYP.  We should be able to hook into it with GN by making a target to execute a shell script.

Patch-set: 3