Update patch set 2
Patch Set 2: Code-Review-1

I ran this and it changed the version name because LASTSWECHANGE was correctly set from the head commit in git, but the versioncode was the same as before. I think we need to see if we can increment the version code too or the update might not apply on top of the previous browser change?

Patch-set: 2
Reviewer: Matt Garnes <9621@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>